util: move zephyr headers into subdir
