+++ include/compiler.h
@@ -32,7 +32,7 @@
 #ifndef _LINUXKPI_LINUX_COMPILER_H_
 #define _LINUXKPI_LINUX_COMPILER_H_
 
-#include "util_macro.h"
+#include "zephyr/sys/util_macro.h"
 
 #ifndef __aligned
 #define __aligned(x) __attribute__((__aligned__(x)))
