+++ include/zephyr/sys/util_internal.h
